사이드바 별점 나와요(가독성이 좀 떨어져용) 댓글 입력창 대충 구현했습니다

+++ src/main/java/kr/tf/spring/model/vo/RestaurantVO.java
@@ -14,5 +14,5 @@ public class RestaurantVO {
 	private String re_phone;
 	private String re_tag;
 	private String re_content;
-	private int re_score;
+	private double re_score;
 }
